Refactor and create waymo_utils.py
